função de add passada para uma pasta separada
